==> sample/02_berger2018_to_csv.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

def peak_file(path):
    '''Peak file data line by line. Enter 'q' to quit.
    '''
    with open(path, 'r') as file:
        stream = file.read()
        lines = stream.splitlines()
        for i, line in enumerate(lines):
            print(f'{i}\t{line}', end='\t')
            if input() == 'q':
                break

path = '../data/berger2018/b18_table1.txt'

peak_file(path)

cols = ['KIC', 'Gaia', 'Teff', 'e_Teff', 'D', '+e_D', '-e_D', 'R', '+e_R', '-e_R', 'AV', 'Evol', 'Bin']
df = pd.read_csv(path, sep='\s+', names=cols, skiprows=32)
df.head()

df.to_csv('../data/berger2018/b18.csv', index=False)
