==> sample/03b_xmatch_serenelli2017_gaia.ipynb <==
# *Gaia*-*Kepler* cross-match

Using [Kepler Gaia Fun!](https://gaia-kepler.fun/) I will cross-match the Serenelli sample with Gaia.

import numpy as np
import pandas as pd
from astropy.io import fits
from astropy.table import Table

kepler_gaia_path = '/Users/ajl573/Data/kepler_dr2_4arcsec.fits'

with fits.open(kepler_gaia_path) as hdul:
    data = hdul[1].data

data.columns

df_kg = Table(data).to_pandas()
df_kg.head()

Flag nearest neighbours

def nearest_neighbours(df, source_id, ang_dist):
    ''' Returns a nearest neighbour flag for each source id '''
    sort = df.sort_values(ang_dist)  # Sort by angular distance
    sort['flag'] = 1  # Flag all as neighbour
    sort.loc[sort[source_id].duplicated(keep='first'), 'flag'] = 0  # Unflag duplicates
    return sort['flag'].sort_index()

df_kg['nearest_neighbour'] = nearest_neighbours(df_kg, 'kepid', 'kepler_gaia_ang_dist')
df_kg.head()

serenelli_path = '../data/serenelli2017/s17_seis.csv'

df_s17 = pd.read_csv(serenelli_path)
df_s17.head()

df_kg.rename(columns={'kepid': 'KIC'}, inplace=True)

df = pd.merge(df_s17['KIC'], df_kg, on='KIC')
print('Length =', len(df))
df.head()

df.to_csv('../data/kepler_gaia/s17_gaia_4arcsec.csv', index=False)
