==> sample/04_xmatch_gaia_2mass.ipynb <==
# *Gaia*-2MASS cross-match

## Depreciated

APOGEE Contains 2MASS photometry, no need to cross-match with 2MASS.

import astropy.units as u
from astropy.table import Table
import pandas as pd
from astroquery.gaia import Gaia
from astroquery.vizier import Vizier

Gaia.login()

path = '../data/serenelli2017_gaia_4arcsec.csv'

df = pd.read_csv(path)
print(len(df))
df.head()

# cols = ['kepid', 'source_id', 'tm_designation']
# table = Table().from_pandas(df[cols])

# job = Gaia.launch_job(
#     "SELECT * \
#     FROM gaiadr2.gaia_source as g, gaiadr2.tmass_best_neighbour as tm\
#     INNER JOIN tap_upload.table AS t \
#         ON g.source_id = t.gaia AND tm.source_id = t.gaia",
#     upload_resource=table,
#     upload_table_name='table'
# )

# results = job.get_results()
# gaia = results.to_pandas()
# print(len(gaia))
# gaia.head()

# gaia['tmass_source_id'] = '2MASS J' + gaia['original_ext_source_id'].str.decode("utf-8")
# gaia['tmass_source_id'].head()

gaia = df

jobs = []
for obj_id in gaia['tm_designation'].to_numpy():
    jobs.append(
        Vizier.query_object(object_name=obj_id, catalog='II/246/out', radius=4.0*u.arcsec)
    )

jobs[:5]

tmass_dfs = []
for job in jobs:
    tmass_dfs.append(job['II/246/out'].to_pandas())

tmass = pd.concat(tmass_dfs)
tmass.head()

# gaia_tmass = gaia.merge(tmass, left_on='original_ext_source_id', right_on='_2MASS')
# gaia_tmass.head()

# total = df.merge(gaia_tmass, left_on='KIC', right_on='kic', suffixes=['', ''])
# total.head()

# total.to_csv('../data/serenelli2017-berger2018-gaia-2mass_all.csv', index=False)

tmass = tmass.reset_index(drop=True)

tmass['tm_designation'] = '2MASS J' + tmass['_2MASS'].str.decode("utf-8")

tmass['tm_designation']

gaia = gaia.rename(columns={'source_id': 'gaia_source_id'})

tmass_gaia = gaia[['KIC', 'tm_designation', 'gaia_source_id', 'nearest_neighbour']].merge(tmass, on='tm_designation')
tmass_gaia.head()

tmass_gaia.to_csv('../data/archive/s17_gaia_4arcsec_2mass.csv', index=False)
