==> sample/05_calculate_asteroseismology.ipynb <==
import numpy as np
import pandas as pd

dr14_path = f'../data/spectroscopy/s17_gaia_4arcsec_apogeeDR14.csv'
dr16_path = f'../data/spectroscopy/s17_gaia_4arcsec_apogeeDR16.csv'

dr14 = pd.read_csv(dr14_path)
dr16 = pd.read_csv(dr16_path)

df_spec = dr14.merge(dr16, on=['KIC', 'nearest_neighbour'], suffixes=('_DR14', '_DR16'))
df_spec.head()

len(df_spec)

seis_path = '../data/serenelli2017/s17_seis.csv'
df_seis = pd.read_csv(seis_path)

cond = df_spec['nearest_neighbour'] == 1
df = df_seis.merge(df_spec[cond], left_on='KIC', right_on='kepid')

df.head()

len(df)

solar_teff = (5777.0, 0.5)
solar_numax = (3090.0, 30.)
solar_dnu = (135.1, 0.1)
solar_logg = (4.4377, 5e-5)

def asteroseismic_logg(df, numax_cols=['numax', 'enumax'], teff_cols=['TEFF', 'TEFF_ERR']):
    '''Returns the asteroseismic logg and propagated error.'''
    
    logg = solar_logg[0] + np.log10(df[numax_cols[0]] / solar_numax[0]) + \
           0.5 * np.log10(df[teff_cols[0]] / solar_teff[0])
    
    logg_var =  ( 
                    1/np.log(10)**2 * ( 
                        (df[numax_cols[1]] / df[numax_cols[0]])**2 +
                        (solar_numax[1] / solar_numax[0])**2 +
                        (0.5 * df[teff_cols[1]] / df[teff_cols[0]])**2 +
                        (0.5 * solar_teff[1] / solar_teff[0])**2 
                    ) + 
                    (solar_logg[1] / solar_logg[0])**2
    )
    
    return logg, np.sqrt(logg_var)

def asteroseismic_radius(df, numax_cols=['numax', 'enumax'], dnu_cols=['Dnu', 'eDnu'], teff_cols=['TEFF', 'TEFF_ERR']):
    '''Returns the asteroseismic logg and propagated error.'''
    
    mass = (df[numax_cols[0]] / solar_numax[0]) * (df[dnu_cols[0]] / solar_dnu[0])**(-2) * \
           (df[teff_cols[0]] / solar_teff[0])**0.5
    
    mass_var = ( (df[numax_cols[1]] / df[numax_cols[0]])**2 +
               (solar_numax[1] / solar_numax[0])**2 +
               (2 * df[teff_cols[1]] / df[teff_cols[0]])**2 +
               (2 * solar_teff[1] / solar_teff[0])**2 +
               (0.5 * df[dnu_cols[1]] / df[dnu_cols[0]])**2 +
               (0.5 * solar_dnu[1] / solar_dnu[0])**2 )
    
    return mass, np.sqrt(mass_var)

def asteroseismic_mass(df, numax_cols=['numax', 'enumax'], dnu_cols=['Dnu', 'eDnu'], teff_cols=['TEFF', 'TEFF_ERR']):
    '''Returns the asteroseismic logg and propagated error.'''
    
    mass = (df[numax_cols[0]] / solar_numax[0])**3 * (df[dnu_cols[0]] / solar_dnu[0])**(-4) * \
           (df[teff_cols[0]] / solar_teff[0])**1.5
    
    mass_var = ( (3 * df[numax_cols[1]] / df[numax_cols[0]])**2 +
               (3 * solar_numax[1] / solar_numax[0])**2 +
               (4 * df[teff_cols[1]] / df[teff_cols[0]])**2 +
               (4 * solar_teff[1] / solar_teff[0])**2 +
               (1.5 * df[dnu_cols[1]] / df[dnu_cols[0]])**2 +
               (1.5 * solar_dnu[1] / solar_dnu[0])**2 )
    
    return mass, np.sqrt(mass_var)

scales = ['SDSS', 'ASPC', 'DR14', 'DR16']
teff_cols = [['Teff', 'eTeff']]*2 + [['TEFF', 'TEFF_ERR']]*2
for scale, teff_col in zip(scales, teff_cols):
    df[f'logg_seis_{scale}'], df[f'logg_seis_err_{scale}'] = asteroseismic_logg(df, 
                                                              teff_cols=[f'{col}_{scale}' for col in teff_col])
    
    df[f'R_seis_{scale}'], df[f'R_seis_err_{scale}'] = asteroseismic_radius(df, 
                                                              teff_cols=[f'{col}_{scale}' for col in teff_col])

    df[f'M_seis_{scale}'], df[f'M_seis_err_{scale}'] = asteroseismic_mass(df, 
                                                              teff_cols=[f'{col}_{scale}' for col in teff_col])

df[['logg_SDSS', 'logg_ASPC', 'logg_seis_SDSS', 'logg_seis_ASPC', 'logg_seis_DR14', 'logg_seis_DR16']]

var = ['logg', 'R', 'M']
cols_to_save = ['KIC']
for v in var:
    for scale in scales:
        cols_to_save.append(f'{v}_seis_{scale}')
        cols_to_save.append( f'{v}_seis_err_{scale}')

cols_to_save

df[cols_to_save].to_csv('../data/asteroseismology/s17_scaling_relations_raw.csv', index=False)

# scales = ['SDSS', 'ASPC']

# for scale in scales:
#     teff_cols = ['_'.join([col, scale]) for col in ['Teff', 'eTeff']]
#     logg_cols = ['_'.join([col, scale]) for col in ['seis_logg', 'seis_logg_err']]
#     mass_cols = ['_'.join([col, scale]) for col in ['seis_mass', 'seis_mass_err']]

#     asteroseismic_logg(df, teff_cols=teff_cols, logg_cols=logg_cols)
#     asteroseismic_mass(df, teff_cols=teff_cols, mass_cols=mass_cols)

# df.plot(*['logg_SDSS', 'seis_logg_SDSS'], kind='scatter', yerr='seis_logg_err_SDSS')
# df.plot(*['logg_ASPC', 'seis_logg_ASPC'], kind='scatter', yerr='seis_logg_err_ASPC')

# df.plot(*['Mass_SDSS', 'seis_mass_SDSS'], kind='scatter', yerr='seis_mass_err_SDSS')
# df.plot(*['Mass_ASPC', 'seis_mass_ASPC'], kind='scatter', yerr='seis_mass_err_ASPC')

# m_lim = (0.85, 1.15)
# feh_lim = (-0.35, 0.35)

# condition = ((df[['Mass_SDSS', 'Mass_ASPC']] > m_lim[0])).all(axis=1) & \
#             ((df[['Mass_SDSS', 'Mass_ASPC']] < m_lim[1])).all(axis=1) & \
#             ((df[['[M/H]_SDSS', '[M/H]_ASPC']] > feh_lim[0])).all(axis=1) & \
#             ((df[['[M/H]_SDSS', '[M/H]_ASPC']] < feh_lim[1])).all(axis=1)

# df_cut = df[condition]
# print(f'Size = {len(df_cut)}')
# df_cut.head()

# %matplotlib notebook
# # ax1 = df_cut.plot(*['Mass_SDSS', 'seis_mass_SDSS'], kind='scatter', 
# #             xerr='+eMas_SDSS', yerr='seis_mass_err_SDSS', alpha=0.5)
# # ax1.plot(np.linspace(0.5, 1.5), np.linspace(0.5, 1.5))

# ax2 = df_cut.plot(*['Mass_ASPC', 'seis_mass_ASPC'], kind='scatter', 
#             xerr='+eMas_SDSS', yerr='seis_mass_err_ASPC', alpha=0.5)

# ax2.plot(np.linspace(0.5, 1.5), np.linspace(0.5, 1.5))


# m_lim = (0.85, 1.15)
# feh_lim = (-0.35, 0.35)

# condition = ((df[['seis_mass_SDSS', 'seis_mass_SDSS']] > m_lim[0])).all(axis=1) & \
#             ((df[['seis_mass_SDSS', 'seis_mass_SDSS']] < m_lim[1])).all(axis=1) & \
#             ((df[['[M/H]_SDSS', '[M/H]_ASPC']] > feh_lim[0])).all(axis=1) & \
#             ((df[['[M/H]_SDSS', '[M/H]_ASPC']] < feh_lim[1])).all(axis=1) & \
#             ((df[['seis_logg_SDSS', 'seis_logg_SDSS']] > 3.6)).all(axis=1)
# df_cut = df[condition]
# print(f'Size = {len(df_cut)}')
# df_cut.head()

df.to_csv('../data/archive/s17-b18-gaia-2mass_seis.csv', index=False)

df[cols_to_save].describe()
