==> sample/09_isoclassify_inputs.ipynb <==
import os
import pandas as pd

df_gseis = pd.read_csv('../data/serenelli2017/s17_seis.csv')
df_scarel = pd.read_csv('../data/asteroseismology/s17_scaling_relations_raw.csv')
df_seis = df_gseis.merge(df_scarel, on='KIC')

scale = 'ASPC'
df_spec = pd.read_csv(f'../data/spectroscopy/s17_gaia_4arcsec_apogeeDR14.csv')
df_spec = df_spec.merge(pd.read_csv(f'../data/serenelli2017/s17_{scale}.csv'), on='KIC')

df_gaia = pd.read_csv('../data/kepler_gaia/s17_gaia_4arcsec.csv')

# Merge on kepler ID and nearest neighbour to account for duplicates
df_comb = df_gaia.merge(df_spec, on=['KIC', 'nearest_neighbour'], suffixes=['_GAIA', f'_{scale}'])
# df_comb = df_gaia.merge(df_spec, on='KIC', suffixes=['_GAIA', f'_{scale}'])

cond = df_comb['nearest_neighbour'] == 1  # Choose only nearest neighbours
df = df_seis.merge(df_comb[cond], on='KIC')
df.head()

df.columns.to_list()

We need to specify the columns for input.

df['parallax']

parallax_offset = 0.05  # mas
df['parallax_cor'] = df['parallax'] + parallax_offset

df['parallax_cor_as'] = df['parallax_cor'] * 1e-3
df['parallax_error_as'] = df['parallax_error'] * 1e-3

df['band'] = 'kmag'

df['dust'] = 'green19'

df['comments'] = f'using DR13 feh and {scale} teff'

df['id_starname'] = 'KIC' + df['KIC'].astype(str)

Dict of columns in master table with their respective columns in the inputs.

cols = {
    'id_starname': 'id_starname', 
    'Teff': 'teff',
    'eTeff': 'teff_err',
    f'logg_seis_{scale}': 'logg',
    f'logg_seis_err_{scale}': 'logg_err',
    '[M/H]': 'feh',
    'e[M/H]': 'feh_err',
    'parallax_cor_as': 'parallax',
    'parallax_error_as': 'parallax_err',
    'J': 'jmag',
    'J_ERR': 'jmag_err',
    'K': 'kmag',
    'K_ERR': 'kmag_err',
    'ra': 'ra',
    'dec': 'dec',
    'band': 'band',
    'dust': 'dust',
    'comments': 'comments',
#     'dmag': 'dmag',
#     'dmag_err': 'dmag_err'
}

inputs = df[cols.keys()]
inputs.head()

iso_inputs = inputs.rename(columns=cols)
iso_inputs.head()

import numpy as np

iso_inputs.replace([-9999.0, -9999.99, -999.99], np.nan, inplace=True)
iso_inputs_dropna = iso_inputs.dropna()

len(iso_inputs_dropna)

Drop negative parallax

iso_inputs_dropna[iso_inputs_dropna['parallax']<0]

iso_inputs_dropnegpar = iso_inputs_dropna[iso_inputs_dropna['parallax']>0]

len(iso_inputs_dropnegpar)

iso_inputs_dropnegpar.to_csv(f'../data/isoclassify/isoclassify_inputs_DR13_{scale}.csv', index=False)

df_iso = pd.read_csv('../data/isoclassify/isoclassify_inputs_DR13_SDSS.csv')
df_iso
